regras de negócio - deletar

+++ services/autor.service.js
@@ -1,4 +1,5 @@
 import AutorRepository from '../repositories/autor.repository.js';
+import VendaRepository from '../repositories/venda.repository.js';
 
 async function createAutor(autor) {
   return await AutorRepository.insertAutor(autor);
@@ -27,9 +28,13 @@ async function updateAutor(autor) {
 
 async function deleteAutor(autorId) {
   await getAutor(autorId);
+
+  const vendas = await VendaRepository.getVendasByAutorId(autorId);
+  if (vendas.length > 0) {
+    throw new Error('O Autor ID informado tem vendas do livro cadastradas.');
+  }
+
   return await AutorRepository.deleteAutor(autorId);
-  // antes de excluir um autor, verificar se existem livros
-  // cadastrados para ele. Caso exista, bloquear a exclusão
 }
 
 export default {
